+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-- Support for scarb workspaces.
+- Support for scarb workspaces
 - Initial version of fuzz testing with randomly generated values
 
 ### Changed